Monitor DOM to find dynamically added <video> tags and remove them onbeforeunload
This way video players added dynamically using e.g. AJAX are detected properly.
Also navigating to a different site would signal to mpris that the player is gone.
So far we would only do that when the tab is closed.

One thing I haven't figured out yet is how to deal with YouTube where the <video>
is apparently re-used, so going from a video to a search result we still think
we have a player. That's not all too terrible, though.