Hook up timeupdate and seeked signals
Timeupdate is called every second when the time changes and stored in the "Position" poperty.
To avoid excess dbus traffic we don't signal this property changing in response to timeupdate,
this is what "Rate" is for so the client can advance the slider on its own.

Media controller reads Position once when it opens and here we used to keep the time of playback
start so the slider would always end up starting again from the beginning.

However, when the user explicitly seeked, signal an explicit position change.