Listen for "waiting" and "canplay" events
Whilst buffering the video (e.g. after a seek operation) set playback state to Paused.
This keeps the seek slider from moving when we're not actually playing something.
Unfortunately MPris doesn't have a "Buffering" state, so that's the best we can do.

If it proves to be annoying (having the tray icon briefly switch to paused while seeking)
we might need another solution (setting "Rate" to 0 perhaps?)