Reduce boilerplate code by allowing to register multiple callbacks in one go
When passing an array as action, a callback is registered for each one of them and
the callback's last argument will be the action that is actually invoked.

This makes registering stuff like play, pause, stop, etc which is more or less
forwarded verbatim a lot less painful to write.