Add some very crude window mapping mechanism
Many of our APIs (quick share, activities, etc) need to get the native window of the browser.
However, the window IDs exposed by the extension API is browser-internal. Our trick is now that whenever
a new window is opened, we open a new tab with some specially crafted page that sets its title to the
browser's window id. In the host we can then detect that and map that to our native window.
After that the tab is closed again.

Yes, that's horrible. At times you can briefly see the tab flashing up, sometimes it also confuses
Chrome's window focus handling, probably when the "New Tab" page takes a bit to load and we area already
closing our mapping tab...