Introduce Settings singleton
This class will eventually marshall the extension's settings in the browser (like which
features the user wants) to the extension host.

It also (so far exclusively) contains information about the environment the host is run in
so we can tell whether we're hooked up to a Firefox or Chrome or other browser.

Might in the future also gain an init callback initially called by the extension on load
to provide further information (like which browser version is actually running etc)