Move browser detection to the extension
Summary:
It's much easier to detect the browser there, so let the extension write
a message with the browser name (mapped to Settings::Environment) to the
native host. Also add support for Vivaldi.

Test Plan:
Installed the extension in Firefox 52, Opera 40 and Vivaldi 1.91,
the icons in krunner appear as expected.

Reviewers: #plasma, broulik, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D6707