Use same file descriptor for notification and reading
Summary:
Otherwise we wait for data on a different descriptor than we read from,
which results in missing notifications if data lands in a buffer.

Test Plan:
No modules were loaded after reloading the extension from the browser,
this is due to the settings changed message not getting received by the host
as it was in quick succession to the setEnvironment message. With this patch
it now works reliably.

Reviewers: #plasma, broulik, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D6714