Expose base64-encoded favicons to the tabsrunner
Summary:
The browser has them in the cache already, so we avoid loading it ourselves,
also we might not even have access to the icons from our context.
Some browsers (Opera does) set the favIconUrl of the tab to the base64 data
already, which is convenient. The base64 URL gets decoded in the krunner
plugin and if it fails or is not available, the browser icon is shown.
If the tab is audible or incognito, that icon is preferred.

Test Plan: Installed in Opera 40, Vivaldi 1.91 and Firefox 52, favicons appear in krunner.

Reviewers: #plasma, broulik, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D6717