Check for HUP/ERR/NVAL conditions on stdin
Summary:
While Qt's event loop checks for POLLHUP, it
treats it exactly like POLLIN and it's not possible
to retrieve that information from Qt. So in the
Connection read handler it needs to be checked for
again.

BUG: 382445

Test Plan:
Before this patch plasma-browser-integration-host
consumed 100% CPU after exiting the browser, 100% reproducible.
Now it quits gracefully.

Reviewers: #plasma, broulik, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D6711