WIP: Show SNI to install browser-side extension
Summary:
PBI comes as two parts, the host and the browser extension which come from different sources.

This daemon notifies the user when the has the host installed, but is yet to install the browser side extension a "broken", but likely setup. It then provides a link to the relevant browser store.

In an attempt to not be too annoying this only appears a few times then removes itself.

I'm away next week, Kai can you commandeer - I don't like the module names.

Reviewers: davidedmundson, apol, mart, #plasma

Subscribers: alexeymin, zzag, hein, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D12698