Make content-script media player code follow user settings
Summary:
If a user disables the mpris plugin there's no point loading anything in
the client page.

Also split the mediaSessions shim into a separate option as it's much
more invasive than the regular mpris controls.

For now the latter is disabled by default. It breaks some sites, we
can't afford to do that and we need to put a fix for that immediately.

We can revisit it later.

Test Plan:
Ran media-sessions API page
With extended mpris on I got thumbnails
With it off I still had the classic video control

Reviewers: fvogt, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D13628