Also intercept creation of video elements
Spotify web player uses a video tag instead of audio to play its music.
Also remove the element after adding it. This ensures the element returned to the caller is not part of the DOM but it is sufficient
for our MutationObserver to see it and keep a reference to it in order to control it.

Differential Revision: https://phabricator.kde.org/D13878