Wait for duration to be known before registering player
A player might emit a play signal before its duration is known.
We want to avoid controlling short sounds (e.g. chat notifications) but I found the duration often is NaN on invocation
when the sound is loaded for the first time.
Defer activating the player until we get a durationchange event for that player

CHANGELOG: Improved algorithm to ignore short sounds, such as chat notifications

Differential Revision: https://phabricator.kde.org/D15724