Recreate interrupted download when it is resumed
When a download is canceled, finishes, or is interrupted (e.g. loss of network connection)
we finish the KJob to remove it from notification area.

However, when a job was interrupted and is then resumed, the download ID is reused by the
browser and we only get sent updates for a job we no longer have.

When we get a change event from interrupted to in_progress just pretend a new download
was created and carry on as normal.

BUG: 401038

Differential Revision: https://phabricator.kde.org/D16888