Only cancel job when an "error" is set and ignore "interrupted"
In Chrome we get state change to "interrupted" and the error in unison whereas Firefox first signals
a state change and then updates the error later.

Previously, we would finish the job as soon as it was interrupted leading to a "unknown error"
any time a download aborted (even if explicitly canceled by the user).

According to documentation the "error" is set in case of an error, so we'll want to cancel the job in this case.

CHANGELOG: Fixed cancelling download in Firefox showing an "unknown error" notification

BUG: 385530
FIXED-IN: 5.14.5

Differential Revision: https://phabricator.kde.org/D17233