Only consider player gone if really no longer part of the visible DOM
We listen for the player being removed from the DOM so we can detect when a player
was removed dynamically when e.g. an overlay div closes.
However, when the player is just reparented elsewhere, we get a DOM add and remove dance,
and then it becomes racy who will be given controls: previous player on another tab or
the same player again when it becomes known again.
To avoid this just ignore player removal when it remains part of the visible page DOM.

Differential Revision: https://phabricator.kde.org/D20582