Handle media session callbacks before passing it on to the browser
The Media Session API standard evolves quite fast, so there's new actions that we might be able to support
that the browser itself doesn't know yet, for instance the "stop" action.
By registering the callback to ourselves before calling into the browser,
we're still able to handle it in the extension even if not supported by the browser.

Differential Revision: https://phabricator.kde.org/D23090