Allow hiding option items depending on available extension and version in the host
We already have somewhat of a versioning system available in the host and it stores and exposes information
about what extension it supports and in what version.
This makes this information available to the options page and allows hiding elements using the
data-requires-extension="foo" and data-requires-extension-version-minimum="2" attributes.

This is for better compatibility when adding new features to the host which is released independently of
the extension. Since the settings are still read and the items still added to the DOM, saving settings will
not touch/change the enabled state of unsupported extensions should you use the extension on multiple
computers with different versions of the host.

Differential Revision: https://phabricator.kde.org/D21113