Enable enhanced media controls by default
Media Sessions API usage becomes more and more common, so it is quite nice to get proper album art.
Also, calling into media session for play/pause when the website provides them is nicer than blatantly calling pause() on the player item.
It also isn't as invasive as the original implementation used to be since it doesn't manipulate the website DOM anymore.

Differential Revision: https://phabricator.kde.org/D23122