Add browser action for status reporting
This adds a tool bar icon for reporting extension status in lieu of the error notification.
When the extension isn't supported or fails to start, the Plasma icon is crossed out.
Especially when synced to a device without plasma-browser-integration, the error notification
that used to show up can be annoying and misleading.
When the extension crashes, an overlay is shown.

Currently, the overlay is disabled when the extension is running properly, but additional controls
such as sharing and media controls blacklist settings could be added to the browser action.
This patch is mostly for the base infrastructure of such a toolbar icon.

Differential Revision: https://phabricator.kde.org/D23126