[Purpose Plugin] Reset pending reply serial when aborting
When the QMenu is closed or the Purpose job finishes, we make sure to send a reply out and reset the pending reply serial.
However, when we determined on the host side that we cannot share the contents,
we send a reply but don't reset the pending reply serial, having every subsequent purpose request
fail until reloading the extension or restarting the browser.

Differential Revision: https://phabricator.kde.org/D25803