[Purpose Plugin] Detect cancelling the prompt more reliably
aboutToHide is emitted before an action is triggered and activeAction() is the action currently hovered.
This means we can't properly tell that the prompt got canceled, when hovering an action and then hitting
Escape to close the menu.
This patch sets a property in response to triggered and then does a delayed check for that in aboutToHide.

See also QTBUG-56761

Differential Revision: https://phabricator.kde.org/D28719