Don't consider player gone when it only got temporarily added by us
We temporarily add players created by JS to the DOM so we can access them from the content-script.
When the player is removed again - at least on Firefox - the MutationObserver notices that and we signal
the player being gone.
This causes media controls for HTML5 Audio to be immediately rescinded when created but not played immediately.

Differential Revision: https://phabricator.kde.org/D28705