Manually merge stored and default settings
One can pass a default Object to storage.get() which will be used for keys that don't exist.
However, it doesn't do that recursively, so when we add a new settings key for a specific module, like the
addToRecentDocuments it will not be present for anyone who already touched the extension settings.

Differential Revision: https://phabricator.kde.org/D28791