Provide light icon for dark mode via theme_icons
I recall we originally didn't do it when Noah added these icons as browserAction.setIcon
wasn't able to override the icon or something like this?
However, it doesn't appear to be the case anymore in more recent versions.

While the setIcons API doesn't support theme icons, only set the custom icon when the extension fails to start.
So let's not jeopardize the normal user experience by this edge case.

BUG: 414544

Differential Revision: https://phabricator.kde.org/D28806