Restore ability to send tel: URLs to the phone
They're not http or https URLs. Add a dedicated handler with appropriate "Call on device" label and icon.
The URL pattern doesn't look like any of the listed examples, but tel://*/* would refuse on grounds that "tel"
isn't a recognized schema but just doing "tel:*" works on both Firefox and Chrome.

Differential Revision: https://phabricator.kde.org/D29691