Split "page" context menu from links
As far as Firefox is concerned, a context menu entry on the page itself has no
"target" and so any `targetUrlPatterns` will cause the menu entry not to show.

This patch splits the "page" context from anything with a link or source, so that
we can set `targetUrlPatterns` for those, and `documentUrlPatterns` for the "page"
menu. This has the added benefit of not allowing to share internal browser pages,
such as "New Tab", while still allowing to send http(s) links on those pages.

While at it, also support the "tab" context in Firefox which adds an entry to the
tab button on the tab bar.