WIP: Use cgroup for detecting which browser we're supposed to mimic
Since the extension host is launched by the browser it will be in the same
cgroup, which we can then use to extract the exact desktop file name.

This way we can also mimic e.g. Chromium when it is packaged differently,
or tell the difference between Firefox and Firefox Nightly, or support
any Chromium-based browser, such as the new Microsoft Edge.