Add browser history runner
Being a DBus runner inside p-b-i has the obvious disadvantage of only working
when the browser is running but I have it open all the time anyway.

It is added as an optional permission as to not cause prompts after upgrading
the extension. Instead, it asks the user to grant the permission either on settings
screen or in KRunner results.

Results are primarily scored by whether the title contains (starts with rated higher),
host contains (+ starts with), path contains (+ starts with) in the same way as is
done in tabs runner, except that the base score is slightly lower. Additionally, the
number of visits to the pages (visited by opening the page, and visited by explicitly
typing into the address bar, the latter of which is scored higher) is taken into account
relative to the result set for added accuracy.