Send along MediaMetadata from content script
It appears the default navigator.mediaSession.metadata object is mirrored into
the content-script context as well. So send it along, if possible, so we at least
get some album art and song info.

Unfortunately, that doesn't seem to give us any way to access or trigger
the registered callbacks. Still better than nothing...