Use libTaskManager for identifying the browser
Leverages its heuristic for finding desktop files and lets us support
pretty much any browser without the need to hardcode a mapping.

The browser still sends us its environment as a fallback. It is also
used so we at least have a baseline of "this is a Firefox derivate" or
"a Chromium browser" since some parts of the host react differently
based on the environment.