[Download Job] Show referrer instead of data URL
A data URL isn't very meaningful to the user, so show the referrer URL
as download source. Also store it in origin URL attribute since a data URL
contains no origin and also isn't very useful as it just contains the data
of the file we've just downloaded.

While at it, optimize `descriptionDirty` to check whether it actually changed
before emitting a description change.