[Downloads] Port to KUIServerV2JobTracker and enable instant reporting
We probably don't want the KIO Widgets job tracker as fallback as
that won't be any more useful than what the browser has buit-in.

Furthermore, set immediate progress reporting, so that the job is shown
right away and always, i.e. even when downloading a small file will show
a "Finished" notification with useful "Open" action.

Finally, while at it get rid of "id" in `DownloadJob` as it was never used.