Check whether removed DOM node was a player
The player might not have been in the DOM in the first place
(e.g. an audio player), so just checking for whether it is in the DOM
at the end looks like a nice optimization but often ends up just
considering the player gone when *any* DOM element got removed.