mprisplugin: Don't skip if size is invalid and take at least one image
The `isValid()` check was supposed to take the first artwork it finds
but a default-constructed `QSize` has zero size, which is a valid size.

Moreover, if we failed to parse the size, we won't be bigger than a
valid size it found before, so the `continue` is redundant and would
cause it not to take the first artwork it found.