Wrap HTMLMediaElement.play instead of Audio constructors
If an audio element is not part of the DOM, the only ways to get it to play
is to use the play method or paused = false. For now just wrap the play method,
which is more likely to be used.

This avoids having to wrap window.Audio and document.createElement, which is
reported to break pages.

BUG: 487303