media-control: Disable on nicovideo.jp to fix video playback
The media session shim, which provides MPRIS media controls, interferes
with the video player script on www.nicovideo.jp. This interference
prevents videos from starting.

This change introduces a workaround to disable the feature for this
specific domain by adding it to the MPRIS_WEBSITE_SETTINGS blocklist
and preventing the problematic script logic from being injected.

This follows the mechanism for problematic sites introduced in !150.

FIXES: 493809